fix [collab]: improve handling of merge commits with renames in code review branches
GitHub does not return a patch for a merge commit, so when a faile is renamed in a merge, the history is effectively lost.

This commit improves the situation by registering an empty history for the file to allow reviewing via the cumulative diff.

This should be enough for the majority of cases, since per-commit review is rare.

IJPL-176633

GitOrigin-RevId: 479785159ebf18577ea3bb5eed656d272a64ba53