refactor [vcs]: Convert VcsUserImpl to Kt and make it @Serializable (IJPL-173924)
There is a breaking change - `VcsUserImpl` was converted to a data class, and the new primary constructor is not converting email to a lower case anymore. Instead, `VcsUserUtil.createUser` should be used.
For this reason the constructor is marked as @Deprecated and will become internal once external usages are converted.

GitOrigin-RevId: e0445da8210e1761ee61840bd0f7faaabe48f6ba