[UI, popups] IJPL-219134 Wayland: resize popups to fit the screen
Some fit-to-screen logic was not invoked for Wayland
before/after showing the popup because it used to rely
on the location as well, which isn't available on Wayland.

Location aside, let's try to at least fit the size into the screen.
The location seems to be fixed automatically anyway
when the window doesn't fit otherwise.
But for the size that automated fixing doesn't seem to always work.
Not clear why, an attempt to reproduce it in a standalone app
failed. But for AbstractPopup, it shows up exactly twice bigger
than the screen, so it would seem to be a scaling issue.
This needs additional investigation, but we can fix it manually for now.

GitOrigin-RevId: 7b93f24dddcfc196f4a18ada1ea695fa7df72456