[UI] IJPL-100455 Maximize dialogs on Windows by double-clicks
The action is already there, the events are received by the JBR,
so all that remains is to invoke the action on a double click.

Note that some windows may have their own double-click
handling. In particular, floating tool windows.
But because both handlers consume the event,
only one handler is invoked (for tool windows it's their
own in the current implementation).

GitOrigin-RevId: ebe80fb913aff359be10a17f0b476d0f06e9483c