[run dashboard] IJPL-209900 Avoid executing syncConfigurations method synchronously inside the topic listener
Now the operation became heavier, we want to schedule it for execution in background and unblock the topic listeners broadcasting flow. The sync itself does not reqire immediate execution, but due to the synchronous nature of the topics it might get executed too early, e.g. on startup or during project importing that negatively impacts the necessary amount of time to finish such crucial operations for no real reason
With the new implementation, no configurations are lost. Previously, due to the volatile boolean flag, we could have ignored some updates and received incorrect results in the end.

IJ-CR-183698

GitOrigin-RevId: a87edcdfd6256c8a93d6085d99c7980ec8c700b2