[terminal] IJPL-206166 Rework the process of showing completion popup in the Reworked Terminal
Previously we mostly relied on the completion triggering logic implemented in the platform (based on `CompletionProgressIndicator`).
It required us to do a lot of hacks, like having our own implementation of `CodeCompletionHandlerBase` with overrides of methods that are designed to be private.
Generally, platform logic was not suitable for the terminal. The main issue we were not able to hack is that it accesses the editor from the background thread under the read action. In the regular editor it is OK, but in the terminal it is not, because we update it in EDT without write action. So, it caused a lot of exceptions mentioned in the ticket.
Now, we implement completion popup triggering logic and suggestions calculation on the terminal side. The only thing we reuse - the completion popup (`LookupImpl`). It gives us much more control and makes the logic much less overloaded.

GitOrigin-RevId: 9f5f6d38cb1994b2eeb64f008dd13d311d8df0ab