[combined-diff]: always restart MUQ on queue (IJPL-211786)
Since it is possible that previous update will "stuck" in MUQ:
E.g., during `flushScheduledUpdates` in `restartTimer` the `scheduledUpdates` becomes cleared.
At the same time, a new update comes in `queue` right after.
The `restartTimer` executed but may return because `scheduledUpdates.isEmpty` or still `isFlushing=true`.
So in the result, the `put` happens in the end and not executed.
Subsequent `queue` of possible "same" updates will be discarded (e.g., because of the same priority) and not trigger execution for already existing updates in the `scheduledUpdates`.

GitOrigin-RevId: ce6fe4ecd1c70cbffbd5d9bfd25f9c932571269a