test [eel]: Fix `EelLocalExecApiTest` TTY esc problems.
WinPTY can do all kinds of weired things: it might move caret, change window title and so on.

We used to filter all ESC commands out, but there are plenty of them.

Not to implement the whole XTERM spec, we delegate it to `JediTerminal`.

The API is cumbersome, so we wrap it with `TTYEmulator`.
It runs background coroutine that provides data to Jedi, and get cleaned output back.

GitOrigin-RevId: da0f6551de3904dd03e6b1574574093daf77d51b