[UI] IJPL-218357 Limit auto size fix attempt frequency
It turns out our automatic window resizing in the case
it becomes too small sometimes does more harm than good.
Something keeps resizing the window back.
We still don't know why it happens, so let's at least
try to limit the frequency. Once per minute sounds OK.
If the first attempt succeeded, then it's unlikely that the next
"true" resize happens within a minute, as it's usually related
to sleep / wake-up or a monitor connecting / disconnecting.
And if it's not a "true" resize, then, whatever it is,
it's better to leave things alone and let the user fix it instead.

GitOrigin-RevId: 6365f977c2fa623350c88bfa546e75c4279952d7