IJPL-221004 ij-light: Use Fast-fail path when checking if path is suitable
Since every path could be potentially be an Eel path, every path should be checked by MRFS providers. However, if some checker is not loaded, it will be loaded by using a `path to class`. That path by itself should be checked to be a valid path. So a cycle extension point creation.

This commit introduces a common for all TCP Eel providers — `/tcp-` to be a fail-fast path

GitOrigin-RevId: b8cc6d4e76841876967c676abd4e3f60595fad80