IJPL-217011 rdct: Restore default settings on frontend and backend simultaneously
The "Restore Default Settings..." action is no longer duplicated in remote development. Instead, a custom implementation of the logic restores the default settings on the frontend and backend simultaneously. This applies to a real remote development environment but not Code With Me where the default settings are restored only on the frontend.

(cherry picked from commit 5858bc3459efb5395112116c935b08c891538235)

GitOrigin-RevId: bc0e0895f4cb0139206209a0a85d80f540f597a1