[UI, Wayland] IJPL-156745 Limit "virtual mouse" movement
Because the protocol tracks the actual mouse movement
and not the cursor movement, the reported location may
run away very far even when we stopped moving the popup.
This leads to an effect when moving the mouse back
does nothing until it "reaches" its previous position.
Meanwhile the cursor is moving, so it looks like the popup
is stuck.

Fix by limiting the reported position movement
when we limit popup movement.
This has an ugly side effect in the case there's
actually space on the screen for the cursor to move
far away from the popup. In this case, it'll immediately
start moving back even before actually reaching the popup.
But it's still way better than not being able to move it at all,
and because we don't know where the screen is,
we can't fix both issues at the same time.

GitOrigin-RevId: 3512978bb236bb0e996985aeb12a714812693f7c