[terminal] IJPL-206166 Reuse the showing lookup in case of completion restart
When prefix length is decreased or there are a lot of completion items, completion process can be restarted to show most actual items.
Previously we were closing the lookup and creating the new one, which caused flickering.
Now we should reuse the existing lookup and perform completion restart without any distractions.

GitOrigin-RevId: ec98f7e4b06cbb0596b40398d6ed6547ac5e7ac8