IJPL-216122 rdct: Propagate account changes from repository to manager
Before these changes, the set of accounts in the AccountManagerBase wasn't updated when the component of the corresponding AccountsRepository was reloaded. This caused, for example, the list of GitHub accounts not being updated after the corresponding settings file was changed during sync.

(cherry picked from commit 742262142496000e3e60b95ddc809f77d5252205)

GitOrigin-RevId: 96fd4e22b5f1f8365c6884214de4a6e33dabf46d