[kotlin][npw][gradle toolchain] KTIJ-36077,IDEA-381447 Use relevant version of foojay plugin for new Gradle 9+ projects
Implement the Gradle to plugins compatibility store (for foojay compatibility).
Also get a foojay version depending on the Gradle version in `GradleSettingsScriptBuilderUtil.kt`.
In Kotlin projects, a relevant version is taken both for new and existing (converted from Java to Kotlin) projects.
In Java projects, a relevant version is taken for new projects.

GitOrigin-RevId: f634ef997a1a8299a022d1477fe66575bbe1477a