[todo] IJPL-208085 Fix `NullPointerException` in the TODO toolwindow when no TODO items
Replaced `List.of()` with `Collections.emptyList()` in `TodoFileNode.getChildrenImpl()` to avoid NullPointException (caused by `List.of()` rejecting `null`).
This exception caused the TODO toolwindow to display "Loading..." forever if there are no TODO items in the file. Also, after removing all the TODO items from the file, the last one removed would stay in the list of the TODOs in the TODO toolwindow.

There was an necessary `isEmpty()` check in split mode, which is now removed.

GitOrigin-RevId: 35dee7399954cf758700938498fb092218f56021