[rdct] IJPL-221336 fix 'fleetClient' occasionally not trying to reconnect
`fleet.rpc.core.DeferredSerializer` is using `rpcCoroutineScope` to run a long-running task.

Prior to the fix, the 'rpcScope' was set to be one from 'consumeAll' in the 'receiver'.
This blocked receiver coroutine from ever finishing and closing the 'requestsChannel'.
As a result, the 'receiver.await()' never finished either and the 'RpcClientDisconnectedException' was not rethrown,
triggering the reconnection cycle.

Fix regression after 338dcda101594c3a735b877cd345a6aabf1319bd where definition of 'this: CoroutineScope' was silently altered.

GitOrigin-RevId: 90ac428813735f2f026ce1e67eb988ed9b8e4f4e