[todo] IJPL-208085 Implement usage of collecting files with TODO items through RPC
`TodoTreeBuilder.collectFiles` is a method that discovers which files in the project contain TODO items (this method is the entry point for building the TODO tree structure).

Its implementation has been modified to use two different approaches, one in split mode and the other in monolith mode. In split mode, `TodoRemoteClient.getFilesWithTodos` is used to query the backend via RPC, avoiding file content access on the frontend. In monolith mode, the original implementation is preserved.

GitOrigin-RevId: 05d08c28ea716012096243893a5d102c82cebec5