[terminal] IJPL-220311 Add the API for executing processes during popup completion suggestions calculation
Originally, the API was designed to execute arbitrary commands right in the shell (`runShellCommand`). But Reworked Terminal doesn't have such an ability - it can only execute the OS processes in the same environment where shell is running.
So, for the Reworked Terminal we provided a custom implementation of `runShellCommand` that split the passed shell command into parts and formed a command line from it. It was rather a hack than a solution, because client code contains commands written this way.
Now is the time to add a new API to be used by completion providers in the right way.

GitOrigin-RevId: 58b64a3529f88fed0988caa8673374d36a5de1b0