[todo] IJPL-208085 Implement usage of counting files with TODO items through RPC
`TodoTreeStructure.getTodoItemCount` is a method that counts the number of TODO items in a file (this method is used to display counts next to file/directory/module names in the TODO tree).

Its implementation has been modified to use two different approaches, one in split mode and the other in monolith mode. In split mode, `TodoRemoteClient.getTodoCount` is used to query the backend via RPC, avoiding file content access on the frontend. In monolith mode, the original implementation is preserved.

GitOrigin-RevId: 35112a4c503c722ec2a4acce34dcad8f5e30f87e