[todo] IJPL-208085 Implement simple version of TODO filter matching through RPC
In split mode, a check that determines if a specific file contains TODO items must be performed on the backend because accessing file content on the frontend would trigger synchronous network calls for `ThinClientVirtualFile`.

The method will be used by `TodoTreeStructure.acceptTodoFilter` to decide whether a file should be included in the TODO tree, allowing the frontend to make filtering decisions without reading remote file content directly.

GitOrigin-RevId: 05397a4979a34cb9fb2e6c6fe3ee53a2f2dea632