[todo] IJPL-208085 Implement usage of TODO filter matching through RPC
`TodoTreeStructure.acceptTodoFilter` is a method that determines whether a file should be included in the TODO tree based on the filter (caled during the building of the tree to filter files).

Its implementation has been modified to use two different approaches, one in split mode and the other in monolith mode. In split mode, `TodoRemoteClient.fileMatchesFilter` is used to query the backend via RPC, avoiding file content access on the frontend. In monolith mode, the original implementation is preserved.

GitOrigin-RevId: 6f053c48b1cc8894e5ed06342120d46b84cb1db6