[todo] IJPL-208085 Implement usage of TODO filter matching in current file scope through RPC
`CurrentFileTodosTreeBuilder.collectFiles` is a method that determines whether the currently opened file should be displayed in the TODO tree (this method is called when the user opens a file in the editor and the TODO panel is in "Current File" scope).

Its implementation has been modified to use two different approaches, one in split mode and the other in monolith mode. In split mode, `TodoRemoteClient.fileMatchesFilter` is used to check the single file via RPC, avoiding file content access on the frontend. In monolith mode, the original implementation is preserved.

`TodoRemoteClient.fileMatchesFilter` is used instead of `TodoRemoteClient.getFilesWithTodos` (and then `contains(psiFile)`) because it more efficiently checks a single file than scanning the entire project in search for the current file.

GitOrigin-RevId: 49b24910eea0acef63b5b3e6e31119405753492b