IJPL-221931 Gateway: restore starting embedded client when running 'Gateway (dev build) + Embedded client' configuration
If the embedded client is started from dev build of Gateway, it needs to return `true` from `PluginManagerCore.isRunningFromSources()`, otherwise after some recent changes in deprecatedPluginDescriptorLoader.kt it fails to load modules. So 'idea.home.path' property should point to the directory with source code, not to 'out/dev-run/Gateway' as it does inside the Gateway process.

GitOrigin-RevId: 42f54fa93117e875e7d4081180ba8a7e7a3a498e