PY-85988 When `cls` in cls()` refers to a dataclass, its synthesized return type is Self
Because the callable type for dataclass callees is provided, the special logic for PySelfType
in `PyCallExpressionHelper.getExplicitResolveResults` didn't apply to dataclasses, so
invoking their constructor as `cls()` in class methods returned a concrete class instance,
not Self. I changed `PyDataclassTypeProvider.getDataclassTypeForClass` to accept a PyType,
instead of a raw PyClass, so PyClassTypeImpl and PySelfType (which is a PyClassType)
should be processed uniformly.

GitOrigin-RevId: 3ef0be3bfaec16283b7b74914e3e2dbee7c683cd