[terminal] IJPL-222625 Fix incorrect PATH when running WSL shell from New Predefined Session popup
Actually, the problem consists of two parts:
1. The fact that we decode `wsl.exe ...` command specified by the client (new predefined session feature or custom user-configured shell path) and start the process remotely.
2. At the same time, shell path seen by `LocalTerminalCustomizer`'s is still `wsl.exe ...`.
So, `PyVirtualEnvTerminalCustomizer` modifies the PATH variable in a way our hacky `TerminalCustomizerLocalPathTranslator` can't handle.
In a result, PATH variable consisting only of file paths from local Windows file system is fed into the remote process. And it broke things.

The current solution is to launch `wsl.exe ...` processes locally. This way we can avoid this problem and also avoid the problem of incorrect parsing of `wsl.exe` parameters (because now we support only `-d`).
Though, the logic of `PyVirtualEnvTerminalCustomizer` should be fixed later as well.

GitOrigin-RevId: 37130d468f2715a8e2d8564d8be3ac79596c6bf3