[js debugger] WEB-75691 Improve UX for 'Copy Value' action
Remove the duplicate Copy JSON action since it works by evaluating the result of calling `JSON.stringfy` on the desired debug field that will be copied. This does not update the UI of the debugger even though e.g. all lazy getters are called. Thus we keep it close to the platform action to copyValue which result is also in JSON.

Co-authored-by: Shumaf Lovpache <Shumaf.Lovpache@jetbrains.com>

Merge-request: IJ-MR-184295
Merged-by: Kiril Panayotov <kiril.panayotov@jetbrains.com>

GitOrigin-RevId: 96fe44373ac9f1f1cdb641ee3cb1748a4447caf4