[terminal] IJPL-219664 Apply hacky fix for missing tool window actions
The problem is that GitHub Copilot plugin is calling `com.intellij.openapi.wm.impl.ToolWindowImpl.getContentManager` from background thread for the Terminal tool window. It triggers initialization of the tool window content and related terminal tool window setup. And `TerminalDockContainer.install` throws an exception because of EDT assertion.
This way, terminal tool window initialization fails to install tab actions, so they are absent in the tool window.
The correct fix should be on the GitHub Copilot side - content manager of the tool window shouldn't be requested from the background thread.
But the temporary fix is to call `TerminalDockContainer.install` last, so it will throw exception only when all previous code is executed.

GitOrigin-RevId: e7c71c000716b8e244bec59389def513f3d3cd35