IJPL-222430 rdct+ij-light: Make `jsch.agentproxy.sshj` and `jsch.agentproxy` libraries content modules
Since JetBrains client requires these libraries as well as the platform core, the libraries package to the `intellij.platform.frontend.split.jar` and that jar is loaded by the core classloader. This commit packages the libraries into the separate jar file, which fixes the problem. Also, this commits bundles the mentioned-above library to the Meta's community build, since it's required for the RD.

Space-RevId: 4560f7998250d3671fa9cd440481bab87e357968

GitOrigin-RevId: a519ed482bb504d4c0d785d566f3af4971d48dcb