PY-86174 Fix behave step definitions detection in nested directories
Well, here's the thing. We clear step registry after dry run (because
we'll reload steps again after running behave for real). Technically we
run behave twice, one for dry run (to get the number of tests and some
context) and then we invoke their runner again.

Here's how behave loads step definitions. It creates a decorator for
each step type, then under the hood it compiles all the python scripts
with steps (in steps directory) and executes the code with definitions
of those decorators as context to execute the script.

When we don't have subfolders, it all works as expected. But here's a
twist with the subfolder. We now have an import for a subfolder script,
but the problem is that sys.modules lives globally for the whole
lifecycle of our runner. So when we execute the same script second time,
we only go through the import statement, we see that our module is
already loaded, and hence don't run the nested steps script, missing the
steps from there.

The fix involves clearing the modules that were loaded during the dry
run, so that we can reimport modules with steps in nested directories.


(cherry picked from commit 9a6e77819aa0ba5d5789049034f2c0e2af4f3b2c)

IJ-MR-185626

GitOrigin-RevId: 8c6870ca8c1b9d24e7d5723b4afcb577d363a15d