refactoring [text-matching]: parameterize matchers with a layout-fixing strategy
In the future, KeyboardLayoutUtil – the current state holder, will be moved to the core platform module. External clients of the library will be responsible for the management state of layout conversion.

This change gives more control to library client. Also, it drops the dependency on ConcurrentHashMap which is absent in kotlin multiplatform. And it drop the dependency on `isLinux` check

GitOrigin-RevId: 627fc2a7dd4f53d1e8967ffebc7de7d712779f21