[terminal] IJPL-220311 Do not use requested working directory as an initial value
Working directory used to start the shell process might be not the actual directory where shell will start.
For example, in case of starting CMD in a directory from WSL, process will start, but CMD will switch the directory to the user's home (because it can't access WSL paths).
So, it is better to rely on the path reported by the shell integration or computed for the process after its start. Until current directory is computed, it will be null.

GitOrigin-RevId: da0c6feee24ef0b005f14913b48a85e7fffa74fe