[kotlin] KTIJ-36778 K2 move: fix missing imports for internal usages
The necessary imports for references inside the moved declarations
might be not added in a case of high GC pressure. This happens if
a switch from AST to stubs happens for the moved file during the
refactoring. The user data attached to the elements inside the file
is lost in this case, including copyable. Since handling of usages
inside the moved declarations is based on user data attached to
references inside, AST unloading leads to incorrect behavior.

GitOrigin-RevId: c6c436add9dfc978064db922205aabfff8bd887f