[terminal] IJPL-220311 Make ShellName obsolete in ShellRuntimeContext
It is not supported in the Reworked Terminal, and it looks like there is no need. So, let's throw `UnsupportedOperationException` if it is called from the Reworked Terminal instead of returning dummy incorrect value.
If we don't find a need in ShellName, then we may drop it once we drop Experimental 2024 Terminal.

GitOrigin-RevId: c6c324b33dc5734131dea46886f22537d50edc5d