[BAZEL-2675] use dynamic_deps_extension without local=True
The authoritative source for our dependency structure is not yet
in all cases files that bazel can read. Therefore, those dynamic
files are included as external repositories where the "fetch" adds
a bzl-file that is then used in BUILD files. These are currently
maked as `local = True`, which certainly gives correct results,
but leads to frequent invalidations and therefore reanalysis,
adding substantial cost to a sync in IntelliJ. As bazel already
recognizes the dependency on the file that is translated to the bzl
file (but not the script that does the translation), we can drop
the `local = True` without putting correctness too much at risk (as
the translation is pretty stable by now) as a temporary workaround
till we fully switch to bazel.


(cherry picked from commit d14a8669d0542cd40db0f2d1c39a63e0c15db9a9)

IJ-CR-185763

GitOrigin-RevId: 48c2cb008f91ce5f825af40685018b78f6912e0b