[todo] IJPL-208085 Use single read action for PSI access in `TodoRemoteApiImpl`
The old, monolith, implementation doesn't explicitly wrap the `PsiTodoSearchHelper.findTodoItems` into a `readAction` block, but it is run under read action, which means that `AsyncTreeModel` is wrapped in read action at some higher level.

With this new implementation for RD, initially it was attempted to minimize read lock duration by splitting into multiple small read actions, but this approach is incorrect. While this means longer read locks that can be canceled and restarted by write actions, this matches the existing behavior of the TODO plugin.

GitOrigin-RevId: 3fc814e141134e719b537118ebf7229baf763be4