[todo] IJPL-208085 Use `Flow` when searching for files with TODOs
Changed `TodoRemoteApi.getFilesWithTodos` to return `Flow<VirtualFileId>` instead of `List<VirtualFileId> for better performance with large projects. Files are now emitted incrementally as they're discovered.

As Java cannot directly work with Kotlin coroutines and Flows, a method is created in the `TodoTreeBuilderCoroutineHelper` that uses `runBlockingCancellable` to bridge between synchronous Java and asynchronous Kotlin.

GitOrigin-RevId: 0beafec234d1fd32729fddfde101d76b6928d744