PY-86367 Avoid showing errors dialog with no messages
The problem is that the dialog doesn't support when there are no
messages in the pool. Using the dialog with empty messages leads to all
sorts of errors, including exceptions being thrown from
invokeOnCompletion coroutine handlers. We end up in such situation
because messages are usually not added in our release builds.

This problem led to OOMs in PyCharm in different scenarios. One of them
is adding an interpreter when there's some error (for example no
.virtualenvs directory exists in user's home). Such scenario effectively
led to an endless loop of handling exceptions without any chance to
stop.

This change explicitly avoids creating a dialog if there are no messages
in the pool.

GitOrigin-RevId: f5da400a8b18bc2dde93c59187c4efbb294821dc