cleanup [textmate]: remove unused dependency from intellij.textmate.backend to intellij.java.indexing.impl (IJPL-224169)
This dependency is not needed for compilation, and there is no corresponding dependency at runtime.
It's also necessary to remove it to enable checks for dependencies with 'provided' scope in PluginDependenciesValidator, see IJPL-224169.

A direct dependency on 'intellij.platform.core.impl' which previously came as a transitive dependency is added.

GitOrigin-RevId: c63ab80a69a4398c8479e91bfe406bafcfe735ca