cleanup [xml]: remove unused dependency from intellij.xml.langInjection to intellij.xpath (IJPL-224169)
This dependency is not needed for compilation, and there is no corresponding dependency at runtime.
It's also necessary to remove it to enable checks for dependencies with 'provided' scope in PluginDependenciesValidator, see IJPL-224169.

An explicit dependency on 'intellij.xpath' is added to 'intellij.appcode.main' to fix AppCodeKeymapsTest.

Reviewed in IJ-MR-186377.

GitOrigin-RevId: 4511563e709feff708feb5ffde3c72c330b6e20b