PY-86422 Do not create service in invokeOnCompletion for inspection
The problem is that if a project is being disposed, getting a service
will throw ProcessCanceledException. And while this is expected in
general, we can't handle it in invokeOnCompletion. This call shouldn't
throw any exceptions, and in this case the exception was caught and
wrapped with CompletionHandlerException. And eventually we got this
exception in our UncaughtExceptionHandler.

To avoid it, we've changed the call to getServiceIfCreated, which will
not throw ProcessCanceledException in such cases. Also note, that it's
not possible not to have uninitialized service in other cases because we
have created it to run the job.

GitOrigin-RevId: c6ff04bdce8d28b95e1b5b78958fda91899164b8