IDEA-383685 Type migration: add missing dependency on intellij.platform.structuralSearch
intellij.java.typeMigration module included in the plugin descriptor module of Java plugin depends on intellij.platform.structuralSearch module, which was converted to a content module, so now it's necessary to specify an explicit dependency on it.

This problem is reported by IdeaUltimatePluginModuleDependenciesTest after restoring checks for dependencies on modules in the core plugin (IJPL-224619).

GitOrigin-RevId: f2a254bc832bd4138dcd7254d627d09e8fff7e28