tests [build scripts]: check that PluginDependenciesValidator works (IJPL-224619)
To avoid problems like IJPL-224619, we need to have a test which verifies that it actually reports problems. So here a test plugin 'TestPluginWithMissingDependencies' is added. Its module contains a dependency in *.iml file which isn't satisfied at runtime, and PluginDependenciesValidatorTest checks that the problem is reported by PluginDependenciesValidator.

Reviewed in IJ-MR-186514.

GitOrigin-RevId: 658216d0cd6f48021f55feb615942a67f4a73e69