IJPL-224619 plugin dependencies tests: don't include content modules in the core classloader's classpath in PluginDependenciesValidator
Initially, it was another to check the parent directory in toPluginLayoutDescription to determine whether a JAR is included in the core classloader's classpath. But after content modules of the core plugin were moved from IDE_HOME/lib/modules to IDE_HOME/lib, this logic stopped working correctly, and non-embedded content modules became included in the core classloader's classpath in the validator.

Now 'loadCorePluginLayout' takes entries only from embedded modules, restoring the correct behavior. A check for such a case is added to PluginDependenciesValidatorTest.

Reviewed in IJ-MR-186514.

GitOrigin-RevId: 366858d832c2b322c5dcbbc4a866cc5347560b4d