LSP-345 fix ClassCastException when formatting a Java file with comments
The formatting code (AbstractJavaBlock#createJavaBlock) was using the deprecated InjectedLanguageUtil#hasInjections,
which internally cast InjectedLanguageManager to InjectedLanguageManagerImpl.
This failed in environments using CoreInjectedLanguageManager instead of InjectedLanguageManagerImpl, like the LSP.

I replaced the usage of InjectedLanguageUtil#hasInjections with `injectedLanguageManager.getInjectedPsiFiles() != null`
(just like the Javadoc of InjectedLanguageUtil#hasInjections suggests).

The root issue is that InjectedLanguageUtilBase#probeElementsUpInner calls InjectedLanguageManagerImpl#getInstanceImpl,
which in an antipattern: dependency on a specific implementation instead of the public interface.

GitOrigin-RevId: 1474f0d2a26800e8d49e0ea2bcd7b8da6d5cd83c