refactor [inlays]: IJPL-230525 override CodeVisionProvider.getDefaultAnchor for Rider
Rider did not receive c1c8859bdff668526a65f83f84e9956e5dc7f0c3 fix as it used to override the method.
This refactoring avoids the unnecessary changes, so we preserve the original behavior.

It is likely that this commit shall be reverted later, and
the VcsAuthorCodeVisionProvider defaults be changed instead.

This fixes 'CodeLensTest.testDefaultAnchorChanges'

GitOrigin-RevId: 142a9ab91ea1eb698434207c838e2b4e88dc2c31