[psi, threading] IJPL-236269: Add specialized version-aware accessors to TreeElement fields for improved performance
To interact with versioned syntax trees, one needs a version, which is stored in a thread-local map. It is expensive from performance point of view to access the map too frequently, especially in tree traversal loops.
So here we are providing new versioned accessors to the tree structure, that can be used to avoid the access to the thread-local map if the version is known in advance.

GitOrigin-RevId: e9d4265449219c7eaa33868f62e658f8f695b904