[ui] fix scaling when copying checkbox icons
When the JBCachingScalableIcon is copied for scaling,
the initial scale value must be propagated from the source icon via
JBScalableIcon#updateContextFrom. CheckboxIcon.WithColor did not have a
constructor that accepts the source icon, so copied instances could end up
with an incorrect size. This commit adds a constructor to CheckboxIcon.WithColor that takes source icon and allows to properly preserve the scale.

IJPL-225566